requirements.txt
requests-oauthlib==1.3.1
python-dotenv==1.0.0
xmltodict==0.13.0
openpyxl==3.1.2